[emitter] hash constant buffer to improve deduplication performance
Summary:

Currently for each constant tensor we need to compare its buffer with
all the emitted buffer before to see if there's a duplicate. This is bad
especially if we have more than one ExecutionPlan in the program. This
PR introduces a map from the hash value to buffer index and hopefully
can largely reduce the deduplication time.

Test Plan:

Reviewers:

Subscribers:

Tasks:

Tags:

ghstack-source-id: 72135bf44fc28223960c1fb204b99dde904495f8
Pull Request resolved: https://github.com/pytorch/executorch/pull/1185